requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.0
python-dotenv>=1.0.1
openai>=1.30.0
pydantic>=2.7.0
aiofiles>=23.2.1
